+++ carboncalc-ui/src/Pages/Home/Home.css
@@ -1,0 +1,44 @@
+.bg-img {
+    background-image: url("https://images.unsplash.com/photo-1607432750402-48f85c94f63a?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2070&q=80");
+    height: 500px;
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: cover;
+    display: grid;
+    place-items: center;
+
+}
+
+.home-container {
+    background-color: #E6EBE0;
+    height:100vh;
+}
+
+.bg-text {
+    color:white;
+    font-weight: bolder;
+    text-align: left;
+    left: 80px;
+    top: 40px;
+    padding-bottom: 0px;
+    margin-bottom: 0px;
+    position: absolute;
+    height: fit-content;
+    margin: auto;
+    font-size: 6.5em;
+    text-shadow: 0 0 55px black;
+    font-family: 'Helvetica Neue';
+}
+
+.bg-detail-text {
+    top: 0px;
+    color:white;
+    position: relative;
+    font-style: normal;
+    font-weight: lighter;
+    font-size: 36px;
+    line-height: 36px;
+    max-width: 800px;
+    
+
+}
