+++ carboncalc-ui/src/Pages/Home/Home.jsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+import './Home.css';
+
+export default function Home() {
+    return (
+        <div className='home-container'>
+            <div className='bg-img'>
+            </div>
+                <div className='bg-text'>Carbon Calc
+                <div className='bg-detail-text'> Greenhouse gases from transportation account for 27% of total US greenhouse gas emissions, check how much carbon your trip takes! </div>
+
+                </div>
+        </div>
+    ); 
+}
