+++ carboncalc-ui/src/theme.ts
@@ -1,0 +1,11 @@
+import { extendTheme } from "@chakra-ui/react"
+
+const theme = extendTheme({
+    styles: {
+      global: () => ({
+        body: {
+          bg: "",
+        },
+      }),
+    },
+  });
